+++ soundcalc/zkvms/zkdtvm/__init__.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-from soundcalc.zkvms.zkvm import zkVM
-
-def load():
-    return zkVM.load_from_toml(Path(__file__).parent / "zkdtvm.toml")